add TensorConv

+++ core/classes/Torch/Class/Tensor/Conv.hs
@@ -4,31 +4,33 @@ import THTypes
 import Foreign
 import Foreign.C.Types
 import Torch.Class.Internal
+import GHC.Int
 
+-- TODO: The first tensor passed in should be the return Tensor -- find out the size and return this instead of mutating it with IO ()
 class TensorConv t where
   {- can't write this out with HsReal, alone
-  validXCorr2Dptr    :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Integer -> Integer -> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Integer -> IO ()
-  validConv2Dptr     :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Integer -> Integer -> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Integer -> IO ()
-  fullXCorr2Dptr     :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Integer -> Integer -> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Integer -> IO ()
-  fullConv2Dptr      :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Integer -> Integer -> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Integer -> IO ()
-  validXCorr2DRevptr :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Integer -> Integer -> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Integer -> IO ()
+  validXCorr2Dptr    :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Int64 -> Int64 -> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Int64 -> IO ()
+  validConv2Dptr     :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Int64 -> Int64 -> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Int64 -> IO ()
+  fullXCorr2Dptr     :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Int64 -> Int64 -> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Int64 -> IO ()
+  fullConv2Dptr      :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Int64 -> Int64 -> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Int64 -> IO ()
+  validXCorr2DRevptr :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Int64 -> Int64 -> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Int64 -> IO ()
   -}
-  conv2DRevger       :: t -> HsReal t-> HsReal t-> t -> t -> Integer -> Integer -> IO ()
-  conv2DRevgerm      :: t -> HsReal t-> HsReal t-> t -> t -> Integer -> Integer -> IO ()
-  conv2Dger          :: t -> HsReal t-> HsReal t-> t -> t -> Integer -> Integer -> Ptr CChar -> Ptr CChar -> IO ()
-  conv2Dmv           :: t -> HsReal t-> HsReal t-> t -> t -> Integer -> Integer -> Ptr CChar -> Ptr CChar -> IO ()
-  conv2Dmm           :: t -> HsReal t-> HsReal t-> t -> t -> Integer -> Integer -> Ptr CChar -> Ptr CChar -> IO ()
-  conv2Dmul          :: t -> HsReal t-> HsReal t-> t -> t -> Integer -> Integer -> Ptr CChar -> Ptr CChar -> IO ()
-  conv2Dcmul         :: t -> HsReal t-> HsReal t-> t -> t -> Integer -> Integer -> Ptr CChar -> Ptr CChar -> IO ()
+  conv2DRevger       :: t -> HsReal t-> HsReal t-> t -> t -> Int64 -> Int64 -> IO ()
+  conv2DRevgerm      :: t -> HsReal t-> HsReal t-> t -> t -> Int64 -> Int64 -> IO ()
+  conv2Dger          :: t -> HsReal t-> HsReal t-> t -> t -> Int64 -> Int64 -> Ptr CChar -> Ptr CChar -> IO ()
+  conv2Dmv           :: t -> HsReal t-> HsReal t-> t -> t -> Int64 -> Int64 -> Ptr CChar -> Ptr CChar -> IO ()
+  conv2Dmm           :: t -> HsReal t-> HsReal t-> t -> t -> Int64 -> Int64 -> Ptr CChar -> Ptr CChar -> IO ()
+  conv2Dmul          :: t -> HsReal t-> HsReal t-> t -> t -> Int64 -> Int64 -> Ptr CChar -> Ptr CChar -> IO ()
+  conv2Dcmul         :: t -> HsReal t-> HsReal t-> t -> t -> Int64 -> Int64 -> Ptr CChar -> Ptr CChar -> IO ()
   {- can't write this out with HsReal, alone
-  validXCorr3Dptr    :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Integer -> Integer -> Integer -> IO ()
-  validConv3Dptr     :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Integer -> Integer -> Integer -> IO ()
-  fullXCorr3Dptr     :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Integer -> Integer -> Integer -> IO ()
-  fullConv3Dptr      :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Integer -> Integer -> Integer -> IO ()
-  validXCorr3DRevptr :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Ptr (HsReal t) -> Integer -> Integer -> Integer -> Integer -> Integer -> Integer -> IO ()
+  validXCorr3Dptr    :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Int64 -> Int64 -> Int64 -> IO ()
+  validConv3Dptr     :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Int64 -> Int64 -> Int64 -> IO ()
+  fullXCorr3Dptr     :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Int64 -> Int64 -> Int64 -> IO ()
+  fullConv3Dptr      :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Int64 -> Int64 -> Int64 -> IO ()
+  validXCorr3DRevptr :: Ptr (HsReal t) -> HsReal t-> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Ptr (HsReal t) -> Int64 -> Int64 -> Int64 -> Int64 -> Int64 -> Int64 -> IO ()
   -}
-  conv3DRevger       :: t -> HsReal t-> HsReal t-> t -> t -> Integer -> Integer -> Integer -> IO ()
-  conv3Dger          :: t -> HsReal t-> HsReal t-> t -> t -> Integer -> Integer -> Integer -> Ptr CChar -> Ptr CChar -> IO ()
-  conv3Dmv           :: t -> HsReal t-> HsReal t-> t -> t -> Integer -> Integer -> Integer -> Ptr CChar -> Ptr CChar -> IO ()
-  conv3Dmul          :: t -> HsReal t-> HsReal t-> t -> t -> Integer -> Integer -> Integer -> Ptr CChar -> Ptr CChar -> IO ()
-  conv3Dcmul         :: t -> HsReal t-> HsReal t-> t -> t -> Integer -> Integer -> Integer -> Ptr CChar -> Ptr CChar -> IO ()
+  conv3DRevger       :: t -> HsReal t-> HsReal t-> t -> t -> Int64 -> Int64 -> Int64 -> IO ()
+  conv3Dger          :: t -> HsReal t-> HsReal t-> t -> t -> Int64 -> Int64 -> Int64 -> Ptr CChar -> Ptr CChar -> IO ()
+  conv3Dmv           :: t -> HsReal t-> HsReal t-> t -> t -> Int64 -> Int64 -> Int64 -> Ptr CChar -> Ptr CChar -> IO ()
+  conv3Dmul          :: t -> HsReal t-> HsReal t-> t -> t -> Int64 -> Int64 -> Int64 -> Ptr CChar -> Ptr CChar -> IO ()
+  conv3Dcmul         :: t -> HsReal t-> HsReal t-> t -> t -> Int64 -> Int64 -> Int64 -> Ptr CChar -> Ptr CChar -> IO ()
